+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFDS3C5ECWK83B8P8WCKP1W5"
+        "01KFDXJJBWZVMBJJY9SETCVS2D"
     )
     r"""The unique identifier of the data source"""